move(package): change package naming snake
