+++ src/main/java/carracing/move_strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package car_racing.move_strategy;
+package carracing.move_strategy;
 
 public interface MoveStrategy {
     int getMovePoint();
