+++ src/test/java/carracing/TrackTest.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
